log-test/requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# LLM and AI
google-generativeai>=0.3.0

# Database and storage
sqlalchemy>=2.0.0

# HTTP requests and async
httpx>=0.25.0
aiohttp>=3.9.0

# Data visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
